# S05T01N02SedoBulpeJordi/S05T01N02SedoBulpeJordi/src/main/resources/api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:9001","basePath":"/","tags":[{"name":"flor-controller","description":"Flor Controller"}],"paths":{"/flor/actualitzarFlor":{"put":{"tags":["flor-controller"],"summary":"Actualitza una Flor","description":"Canvia els atributs d'una flor, segons el seu id.","operationId":"actualitzarFlorUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"flor","description":"flor","required":true,"schema":{"$ref":"#/definitions/Flor"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/flor/afegirFlor":{"post":{"tags":["flor-controller"],"summary":"Afegir una flor","description":"Se li passa un body JSON amb els atributs de l'Entitat excepte el id, que es autoincrementat.","operationId":"afegirFlorUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"flor","description":"flor","required":true,"schema":{"$ref":"#/definitions/Flor"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/flor/esborrarFlor/{id}":{"delete":{"tags":["flor-controller"],"summary":"Esborra una Flor","description":"Elimina un registre de flor de la BBDD","operationId":"esborrarFlorUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/flor/mostrar/{id}":{"get":{"tags":["flor-controller"],"summary":"Mostra una flor","description":"Mostra una flor de la base de dades en la versió DTO que inclou un atribut addicional respecte de l'Entitat: TipusFlor","operationId":"mostrarUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/flor/mostrarTotes":{"get":{"tags":["flor-controller"],"summary":"Mostra Flors","description":"Mostra una llista amb totes les flors de la base de dades. DTO.","operationId":"mostrarFlorsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"Flor":{"type":"object","required":["nomFlor","paisFlor","pk_FlorID"],"properties":{"nomFlor":{"type":"string","description":"Nom de la flor"},"paisFlor":{"type":"string","description":"País de procedència de la flor"},"pk_FlorID":{"type":"integer","format":"int32","description":"Id de la flor - autoincrementat"}},"title":"Flor"}}}